Switch from storing parameters with pytorch to zarr using tensorstore.
