+++ bin/git-cml-filter
@@ -11,7 +11,7 @@ import logging
 import hashlib
 from collections import defaultdict, OrderedDict
 
-from git_cml import git_utils, checkpoints, params
+from git_cml import git_utils, checkpoints, params, file_io
 
 logging.basicConfig(
     level=logging.DEBUG,
@@ -46,14 +46,14 @@ def clean(args):
     model_dir = git_utils.create_git_cml_model_dir(repo, args.file)
 
     model_checkpoint = checkpoints.PyTorchCheckpoint.from_file(sys.stdin.buffer)
-    staged_file_contents = OrderedDict({"model_dir": model_dir})
+    staged_file_contents = OrderedDict({"model_dir": os.path.relpath(model_dir, repo.working_dir)})
     for param, keys in checkpoints.iterate_dict_leaves(model_checkpoint):
         param_name = "/".join(keys)
         staged_file_contents[f"{param_name} shape"] = params.get_shape_str(param)
         staged_file_contents[f"{param_name} dtype"] = params.get_dtype_str(param)
         staged_file_contents[f"{param_name} hash"] = params.get_hash(param)
 
-    git_utils.write_staged_file(sys.stdout, staged_file_contents)
+    file_io.write_staged_file(sys.stdout, staged_file_contents)
 
 
 def smudge(args):
@@ -63,15 +63,16 @@ def smudge(args):
     logging.debug(f"Running smudge filter on {args.file}")
 
     repo = git_utils.get_git_repo()
-    staged_file = git_utils.load_staged_file(sys.stdin)
+    staged_file = file_io.load_staged_file(sys.stdin)
 
     model_dict = defaultdict(dict)
-    for file, keys in checkpoints.iterate_dir_leaves(staged_file["model_dir"]):
+    for leaf_dir, keys in checkpoints.iterate_dir_leaves(os.path.abspath(staged_file["model_dir"])):
+        param_file = os.path.join(leaf_dir, 'params')
         logging.debug(f"Populating model parameter {'/'.join(keys)}")
         d = model_dict
         for k in keys[:-1]:
             d = d[k]
-        d[keys[-1]] = git_utils.load_tracked_file(file)
+        d[keys[-1]] = file_io.load_tracked_file(param_file)
 
     model_checkpoint = checkpoints.PyTorchCheckpoint(model_dict)
     model_checkpoint.save(sys.stdout.buffer)
